# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Cost Control Tower — Ingestion API
  version: "1.0.0"
  license:
    name: Internal use only — not for external distribution
  description: |
    ## Getting started

    This API lets an internal HCLTech application write its own AI usage and
    outcome events into the same Cost Control Tower tables Product Studio
    already uses, so it can be traced, costed, and governed the same way.

    **Registration is not self-service.** Contact the platform operator to
    register your app and receive a credential — registration and credential
    issuance are direct SQL/RPC, run manually, not exposed over HTTP
    (see the spec's Section 3/4 for why).

    **Credential handling.** Your API key is shown exactly once, at issuance
    time. Move it into your own team's secrets manager immediately — it is
    never stored in plaintext or recoverable after that point. Every request
    authenticates with `Authorization: Bearer <api_key>`; `company_id` and
    `app_id` are resolved from the credential server-side and are never
    accepted as request fields.

    ```bash
    curl -X POST https://<ingestion-host>/v1/usage-events \
      -H "Authorization: Bearer <your_api_key>" \
      -H "Content-Type: application/json" \
      -d '{
        "client_call_id": "a1b2c3d4-...",
        "user_role_at_call": "engineer",
        "caller": "my-app-feature-x",
        "requested_model": "gpt-4o",
        "status": "success",
...
      }'
    ```

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/outcomes:
    post:
      operationId: createOutcome
      tags: [outcomes]
      summary: Start an outcome
      description: |
        Always a plain create — never silently hands back an existing
        resource on its own. `client_outcome_id` is your own idempotency
        key (same role as `client_call_id` on usage events) and is
        mandatory, so retry-safety isn't something you have to opt into.
        Server sets `status='in_progress'`, `started_at`, `last_activity_at`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome_type_id, client_outcome_id, session_id]
              properties:
                outcome_type_id: { type: string, description: "Must already be registered via POST /v1/outcome-types" }
                client_outcome_id: { type: string, format: uuid }
                session_id: { type: string, format: uuid }
                user_id: { type: string, format: uuid, nullable: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  outcome_id: { type: string, format: uuid }
                  deduplicated:
                    type: boolean
                    description: "true if this client_outcome_id was already recorded — the returned outcome_id is the original, not a new one."
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/outcomes/{id}:
    patch:
      operationId: completeOutcome
      tags: [outcomes]
      summary: Mark an outcome completed
      description: |
        Ownership-checked: the outcome must belong to the credential-resolved
        `(company_id, app_id)` and currently be `in_progress`. Returns 404
        without distinguishing "doesn't exist," "belongs to someone else,"
        or "already completed" — a 404 never confirms which UUIDs are real
        for a caller that doesn't own them.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [completed] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  outcome_id: { type: string, format: uuid }
                  status: { type: string, enum: [completed] }
                  completed_at: { type: string, format: date-time }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/outcome-types:
    post:
      operationId: registerOutcomeType
      tags: [outcome-types]
      summary: Register (or confirm) an outcome type
      description: |
        Idempotent upsert — but "idempotent" only means safe to resubmit
        IDENTICAL values. Re-registering an existing `outcome_type_id` with
        a different `costing_method` or `unit_label` is rejected with 409,
        not silently applied. Register a new `outcome_type_id` instead if
        your taxonomy genuinely changes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutcomeTypeInput'
      responses:
        '200':
          description: Created, or an identical resubmission (no-op).
          content:
            application/json:
              schema:
                type: object
                properties:
                  outcome_type_id: { type: string }
                  status: { type: string, enum: [created, unchanged] }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: outcome_type_id already registered with a different costing_method or unit_label.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
    get:
      operationId: listOutcomeTypes
      tags: [outcome-types]
      summary: List your own registered outcome types
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  outcome_types:
                    type: array
                    items: { $ref: '#/components/schemas/OutcomeTypeInput' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/company-apps/me:
    get:
      operationId: getCredentialStatus
      tags: [company-apps]
      summary: Self-check on your own credential
      description: Never returns credential_hash.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  company_id: { type: string, format: uuid }
                  app_id: { type: string }
                  is_active: { type: boolean }
                  credential_created_at: { type: string, format: date-time }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The API key issued to your app, as a Bearer token.
  responses:
    BadRequest:
      description: Malformed payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Unauthorized:
      description: Missing, invalid, or inactive credential
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    NotFound:
      description: Referenced resource doesn't exist (or doesn't belong to the caller)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            type: { type: string }
            message: { type: string }
...
    OutcomeTypeInput:
      type: object
      required: [outcome_type_id, name, description, canvas, costing_method, unit_label]
      properties:
        outcome_type_id: { type: string, description: "Your own taxonomy key" }
        name: { type: string }
        description: { type: string }
        canvas: { type: string, description: "General feature-or-origin label for non-Product-Studio apps" }
        costing_method: { type: string, enum: [session_sum, yield_ratio] }
        unit_label: { type: string, description: "Noun matching units_generated" }
        abandonment_window_hrs: { type: integer, nullable: true, description: "Hours of inactivity before auto-flagged abandoned" }